Removes TT bug & adds string representation
- Bug was that previously attempting to overwrite a transposition for a
depth which already existed in an entry would throw an error. This never
happened before because the result would just be queried from the table
and not updated. But now this can be happen due to multiple node types
(i.e. an all node can be overwritten by an exact node).