Adds new checkmate/premate extraction
- Previous extraction method only allowed for ~20k checkmates to be
extracted from Kingbase collection games. This is because we could only
store ~ 1 million single game pages in a single directory, which
contained those ~20k checkmates. This is due to OS limits on the number
of files a directory can contain.
- This was resolved by not requiring single games to be extracted before
finding checkmates, and instead iterating through the PGN collections
directly.
- ~50k checkmates were extracted from the Kingbase 2017 and 2018 game
collections.